requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
python-dotenv>=1.0.0
langchain-community>=0.0.16
langchain-core>=0.1.17
pymongo>=4.6.0
influxdb-client>=1.36.0
